requirements.txt
telegram==21.7
fastai==2.7.18
torch==2.5.1+cu121
numpy==1.26.4
opencv-python==4.10.0
Pillow==11.0.0
nest-asyncio

